# swagger/v1/todo.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "v1/todo.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "TodoService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/bookings": {
      "get": {
        "operationId": "TodoService_Sample",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/TodoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "TodoService"
        ]
      },
      "post": {
        "operationId": "TodoService_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/TodoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TodoRequest"
            }
          }
        ],
        "tags": [
          "TodoService"
        ]
      }
    }
  },
  "definitions": {
    "Todo": {
      "type": "object",
      "properties": {
        "id": {
          "$ref": "#/definitions/UUID"
        },
        "name": {
          "type": "string"
        },
        "done": {
          "type": "boolean"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "TodoList": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Todo"
          }
        }
      }
    },
    "TodoRequest": {
      "type": "object",
      "properties": {
        "todo": {
          "$ref": "#/definitions/Todo"
        }
      }
    },
    "TodoResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "UUID": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}